Updated bug report template
Modified the GitHub issue template for bug reports. Added "triage" to the labels for better categorization. Removed 'None' from module options as it was redundant. Expanded the description for 'Link to Reproduction' section, providing more detailed instructions on how to provide a useful reproduction of the bug, including minimal code snippets, repository links and live examples.