Updated component template handling and cleaned up Vue code
In this update, we've added a new method to check if a component's template can be run separately. This is determined by the settings of the component itself. If it cannot be run separately, the method will return false.

We've also made changes to some Vue files, cleaning up unnecessary lines and improving readability. A HTML template file was deleted as part of these changes.

Lastly, there were modifications in binary files and logs which are not directly visible but contribute to overall performance improvements.